feat(interface): use interceptor to pass access token
