+++ interface/components/RequireAuth.tsx
@@ -5,14 +5,19 @@ import { useNavigate, useLocation } from "react-router-dom";
 
 import { UserContext } from "../context/userContext";
 
-import { useApi } from "../hooks/useApi";
+import { headerInterceptor, unAuthorizedInterceptor } from "../utils/interceptors";
 
 const RequireAuth = ({ children }: { children: React.ReactNode }) => {
   const navigate = useNavigate();
   const { pathname } = useLocation();
   const { ready, connected, logout } = useContext(UserContext);
 
-  useApi(logout);
+  unAuthorizedInterceptor(logout);
+
+  const token = localStorage.getItem("access-token");
+  if (token) {
+    headerInterceptor(token);
+  }
 
   // If the user is not connected, redirect to the home page
   useEffect(() => {
