--- interface/hooks/useApi.tsx
@@ -1,22 +1,0 @@
-import { useEffect } from "react";
-import axios, { AxiosError } from "axios";
-
-export const useApi = (logout: () => void) => {
-  useEffect(() => {
-    const interceptor = axios.interceptors.response.use(
-      (response) => {
-        return response;
-      },
-      (error: AxiosError) => {
-        if (error.response?.status === 401) {
-          logout();
-        }
-        return Promise.reject(error);
-      }
-    );
-
-    return () => {
-      axios.interceptors.response.eject(interceptor);
-    };
-  }, []);
-};
